pyproject.toml
[tool.poetry]
name = "pycascades-cms"
version = "0.1.0"
description = ""
authors = ["Seb Vetter <seb@pycascades.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.0"
wagtail = "^2.11.3"
wagtail-bakery = "^0.5.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
pyjade = "^4.0.0"
django-assets = "^2.0"
django-storages = "^1.11.1"
sentry-sdk = "^0.19.5"
cssmin = "^0.2.0"
jsmin = "^2.2.2"
pynetlify = "^0.1.1"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
